requirements.txt
PyOpenGL
PyOpenGL_accelerate
numpy

